services/api/requirements.txt
fastapi
uvicorn[standard]
websockets
playwright
sqlalchemy
pydantic
pydantic-settings
httpx
croniter
beautifulsoup4
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.2.1
